wait for presentation upload complete before sending notification about new current presentation
