+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/presentationpod/PresentationConversionCompletedSysPubMsgHdlr.scala
@@ -2,10 +2,11 @@ package org.bigbluebutton.core.apps.presentationpod
 
 import org.bigbluebutton.common2.msgs._
 import org.bigbluebutton.core.bus.MessageBus
-import org.bigbluebutton.core.db.PresPresentationDAO
+import org.bigbluebutton.core.db.{NotificationDAO, PresPresentationDAO}
 import org.bigbluebutton.core.domain.MeetingState2x
 import org.bigbluebutton.core.models.PresentationInPod
 import org.bigbluebutton.core.running.LiveMeeting
+import org.bigbluebutton.core2.message.senders.MsgBuilder
 
 trait PresentationConversionCompletedSysPubMsgHdlr {
   this: PresentationPodHdlrs =>
@@ -55,6 +56,18 @@ trait PresentationConversionCompletedSysPubMsgHdlr {
       pods = pods.addPresentationToPod(pod.id, presWithConvertedName)
 
       PresPresentationDAO.updatePages(presWithConvertedName)
+      if(pres.current) {
+        val notifyEvent = MsgBuilder.buildNotifyAllInMeetingEvtMsg(
+          meetingId,
+          "info",
+          "presentation",
+          "app.presentation.newPresentationActiveNotification",
+          "Notification when a new presentation is set as current",
+          Vector(s"${pres.name}")
+        )
+        NotificationDAO.insert(notifyEvent)
+      }
+
       state.update(pods)
     }
