+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/presentationpod/PresentationPageConversionStartedSysMsgHdlr.scala
@@ -55,7 +55,7 @@ trait PresentationPageConversionStartedSysMsgHdlr {
       var pods = state.presentationPodManager.addPod(pod)
       pods = pods.addPresentationToPod(pod.id, pres)
       if (msg.body.current) {
-        pods = pods.setCurrentPresentation(liveMeeting.props.meetingProp.intId, pod.id, pres)
+        pods = pods.setCurrentPresentation(pod.id, pres)
       }
 
       state.update(pods)
